+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/JibPluginConfiguration.java
@@ -52,9 +52,6 @@ abstract class JibPluginConfiguration extends AbstractMojo {
 
   @Nullable @Parameter String mainClass;
 
-  @Parameter(defaultValue = "true", required = true)
-  boolean enableReproducibleBuilds;
-
   @Nullable
   @Parameter(defaultValue = "Docker", required = true)
   String imageFormat;